Implemented Separate Classes
I separated the classes into their own files and added a new class. I streamlined the call to Form 2, and also enabled the dates to be displayed in Form 2's read-only text boxes. Outside of troubleshooting, the rest of my focus will be on implementing a list, which is a data structure that internally uses arrays. In addition, I'll be adding some exception handling, and events!