Added Exceptions and Events
I've now altered the CarRegistration class to enable defined exceptions, and to clean up the code a little further. I've also added events and removed the Find Available button, because it is no longer necessary with eventhandler. All that's left is to change CarDatabase to utilize a better data structure and ensure it utilizes proper exception handling as well. Primarily if a database.txt doesn't exist, and if it reaches a null variable.