Updated Form1 and the various classes with error checking
I've added a lot of exception handling to the classes I implemented, and to form1. I'm currently in the process of adding one more exception to a class, and imlementing a list. This change is mainly for my current progress, and to remove all of the excess files that were generated in the previous commit. It should still have the bulk of what was added.